refactor: split version file into two
